server: Add /id GET handler
This handler returns the user id from the authorization token.
This isn't specified in the requirements but is useful for testing
the authorization logic.

Adds the function to check if a user is logged in and to expire or
refresh the log in, for future use when transfers is implemented.